requirements.txt
requests
python-dotenv
pokemontcgsdk
CurrencyConverter